requirements.txt
pandas
datasets
transformers[torch]
librosa
evaluate
jiwer
huggingface_hub
webdataset
tensorboardX
indic_nlp_library
accelerate
wandb